# @FIR-770 - LLama.cpp: Adding Perf Status for all GGML Operation